[REFACTOR] Update tests files; make tests working
